+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_index/TestDVUpdateBackwardsCompatibility.java
@@ -54,7 +54,7 @@ public class TestDVUpdateBackwardsCompatibility extends BackwardsCompatibilityTe
   @ParametersFactory(argumentFormatting = "Lucene-Version:%1$s; Pattern: %2$s")
   public static Iterable<Object[]> testVersionsFactory() {
     List<Object[]> params = new ArrayList<>();
-    // TODO: why are we only testing one version here?
+    // NOCOMMIT: why are we only testing one version here?
     params.add(new Object[] {Version.LUCENE_9_0_0, "dvupdates.%1$s.zip"});
     return params;
   }
